Robustly map frames for the kernel
Before, we used a fragile system to map the frames required by the kernel. Here, we change this out for
a more robust system that handles structures that overlap in page boundaries, what their flags should be
and finally correctly map frames allocated to GRUB modules. This fixes our test of executing code loaded
as a GRUB module.