Use newtype pattern to make memory addresses safer
Replace type aliases `PhysicalAddress` and `VirtualAddress` with newtypes to stop accidental conversion between physical and virtual memory addresses, which was possible before since both were just
`usize`s. Converting between the two is now a pain in the ass, which is by design, to make the developer think about whether they're actually trying to do that.