Migrate x86_64 into a separate crate, merge multiboot2 into it
I thought it might work better if `x86_64` was a separate crate, rather than a module, so `multiboot2` could depend on it and use things like PhysicalAddress and VirtualAddress. However, this would've created a circular dependency so I had to merge `multiboot2` into the `x86_64` crate. At this rate, it would've probably been simpler to just merge `multiboot2` into the main kernel crate, but there we go.