[boot x86_64] Change bootloader allocator to pre-allocate frames
Change the frame allocator used to allocate memory for the page tables in
the bootloader to pre-allocate a bunch of frames, then allocate out of
them. This means it doesn't allocate when mapping stuff, so we run less of
a risk of invalidating the key of the memory map between getting it and
calling ExitBootServices.