Clean up linker script and add custom program headers
We're going to move to loading the kernel ELF properly and so need program
headers with the correct permissions (I hadn't realised how easy it is to
emit custom ones before, so stuck with sections).

This also allows us to clean up how we lay out sections because we don't
care about them anymore.