[BREAKS BUILD] Rewrite x86_64 paging module
We're moving to a simpler technique where we just map the entirity of
physical memory into the kernel virtual address space. This makes the
recursive mapping redundant, which actually makes the page-table code much
simpler, as both the bootloader's identity mapping, and the kernel's
"physical memory is all there, just at this offset" are basically the same.
This should also make handling other task's page tables much easier,
instead of using that awful `with` trick.

The new page table code also supports huge 2MiB pages properly, and it
should be trivial to add support for 1GiB pages later if we decide we want
them.