Install per-CPU data and make TSS per-CPU
We use the GS segment to point to per-CPU data so it can easily be accessed
from around the kernel. On x86_64, GS is one of the few segments that can
have a non-zero base, but it has to be managed through an MSR, IA32_GS_BASE,
rather than through GDT entries like on x86.

We also move the TSS into the per-CPU data, because we'll need one per CPU
when we begin scheduling tasks on more than one core.

This commit may also contain residual changes due to the upcoming work on
scheduling.