Split per-CPU data into arch-common and arch-specific parts
We need to do this because we want to be able to access per-CPU data, such
as the scheduler, from the arch-common parts of the kernel, such as the
syscall handlers, but we don't want to expose the arch-specific per-CPU
data types from the arch modules. We solve this by defining the structure
of the common parts, and exposing it in a way that can be extended to other
architectures later.