Fix long-standing issue with the buddy allocator
Sigh. I have no idea how we got this far without hitting this. After many
days of looking through the broken context switching code, I eventually
realised we were allocating the same physical frames multiple times - this
led to screwed up page tables, kernel and user stacks, and an all-round bad
time. The culprit was an issue with the buddy_of function. In the future,
we should definitely expand the allocator unit testing, both to retain my
sanity and to be able to expand the allocator scheme in the future with
confidence.

This fixes the buddy_of function and adds some basic unit tests for it.
Also rewrites some of the module documentation.