Get basic context switching + scheduling working!!!
This is obviously far from a real context-switching system (it only
supports storing the general register state, for a start), but is enough
to make our kernel a real™ multitasking system (in theory). This is a
little messy still and touches a lot of parts of the kernel that could use
some TLC, but does appear to be robust!

This also changes the registers we use on x86_64 for system calls, as the
new scheme requires less faffing about.