Add an InitGuard type for initializing globals
This type is meant to safely encapsulate the idea that a global can be
initialised at some point, and then is safe to use. I'm sure there must
be an existing way to represent this that doesn't involve using a `static
mut` but I haven't come across it.

At first glance, this might seem similar to `spin::Once`, but the semantics
(even if the underlying mechanism is very similar) are slightly different -
`Once` is meant for values that should only be calculated once, but can be
initialized at any point. `InitGuard` is meant for things that need to be
initialized at a key point during bring-up - it's straight up initialization,
as opposed to memoization.