Use InitGuard to make common and arch data global
I'm not sure about this approach - I'm worried this'll end up being very
messy (especially if these structs end up with a tonne of data in) but I
think this is just a common problem for kernels. In any case, I think this
is a good avenue to explore and certainly nicer than some of the other
designs I've seen in Rust kernels.