Make mapping of boot info more robust (and fix it)
When the size of the boot info exceeded a single page, we forgot to update
the mapping logic to map it into the kernel address space to map more than
the single page. This meant we were actually reading out of the local APIC
config space when reading the end of the boot info (oops).