System call to map a MemoryObject into an AddressSpace
This is just a basic version at the moment, and needs to be fleshed out
with extra sanity checking and permission checking in the future. This is
also the first time we've had to call into real existing kernel
functionality from a syscall handler, so we had to rearrange some stuff to
do with the object map and such.