Clear the framebuffer in userspace!
This just writes a whole bunch of magenta to the framebuffer, but from
userspace this time! The userspace code is very fragile because we don't
yet have a way to tell the driver about the framebuffer (e.g. address,
resolution, format etc.), but it works for the framebuffer used in QEMU for
now.