Fix potential bug in InitGuard
Not sure I 100% understand enough about atomics (and so probably shouldn't
be writing sync primitives like this...) but I think the old version had
a potential problem in which one CPU checks that the state is STATE_UNINIT,
but then another CPU changes it to STATE_INITIALIZING. Both CPUs then race
through the initialization, which obviously isn't what we want. This new
version atomically sees if the state is STATE_UNINIT and changes it to
STATE_INITIALIZING if it is, removing that race.