Develop drivers under a single workspace
This will hopefully dramatically reduce build time, as it was before
building the entire sysroot for the custom target for every single driver
(this would obviously not be a problem while we've only got one driver,
but seems like a good change to make now).

Also make `make clean` not complain if pebble.img is not there when you
try to delete it.