Add infrastructure for blocking tasks
This changes the scheduler to be able to "block" tasks if they can't make
any progress until a specific event happens (e.g. Mailbox receives mail,
or to sleep for a set amount of time etc.). It currently doesn't unblock
tasks - this functionality will be added as reasons for blocking are
fleshed out.