Make canonicalizing VirtualAddresses easier
It became a bit of a pain to canonicalize virtual addresses, and led to a
bunch of `unwrap`s being spattered around everywhere. Now, we just always
canonicalize addresses when they're created, and swallow the tiny
performance cost for now.

This commit also includes a bunch of changes needed around the kernel re
this change, but is bound to have missed some.
is bound to