Rename `Mapper` trait to `PageTable`
`Mapper` is from a by-gone era where you had to separately create a mapper
from a set of page tables because you needed it to be installed into the
recursive entry of the active tables. This method is far better and doesn't
require a separate structure and so it seems weird to keep this naming in
hindsight.