Introduce generalized PerCpu type
This is still kinda messy and I'm not sure this is how we'll stick with it,
but it's good to get this preliminary version down. Basically, the HAL
needs to supply a method for registering per-cpu data, and it also needs to
control the overall structure of the per-cpu structure (because it's nice
to be able to use offsets directly in assembly etc., but we also need to
store common data (this isn't just an organisational thing - the HAL
doesn't have access to the kernel types, such as the scheduler).