Add better versions of the pin_utils macros to pebble_util
This adds versions of the macros that can include visibility modifiers so
you can add projecting functions that are visible to the outside world too
(I didn't put too much thought into why pin_utils doesn't do this -
hopefully it's not because it's a safety issue and I don't see this message
in git blame someday).

Also drinks the coolaid and adds the weird `core_reexport` trick to
pebble_utils.