kernel: break all of the system calls for now
These all still use the old kernel object framework, but this is part of a
huge rework where we introduce handles and stuff, and so I don't want to
do this now. Instead, we've got a few working to prove that we haven't
broken them, and will rework the whole system soon.