efiloader: add janky memory map hack
Basically, we've seen an issue on a few UEFI implementations where we
allocate a memory map of the size it tell us, then it tells us the buffer
is too small for the memory size. We now utilize the idea that, since we
might have to allocate more memory anyway, we might as well make it
available to GetMemoryMap. In the majority of cases, this should prevent
us from hitting the issue by fluke...