pebble_util: bring in-tree downcast-rs and make needed modifications
This temporarily brings the downcast-rs crate in-tree and makes changes to
it to allow it to be used from no_std contexts. In due course, I either
want to either upstream these, or write a better library to do this in a
nicer way, but that can happy at some other time.

I've added some comments in-line about questions about making the library
nicer to use / improving the implementation.