kernel: allow downcasting of KernelObject trait objects
This allows us to safely turn an Arc<dyn KernelObject> into the actual type
of kernel object it is. This allows us to take a handle and actually
operate on the kernel object.

This requires making a few things in the HAL `Send`-bound. This is a bit
leaky but I think is needed.

I think I want to re-write the actual downcasting library and so this might
change in the future but I think it's good for now as a proof-of-concept.