kernel: start of platforms and add hal_arm64 plumbing
We're starting work on our first ARM64 platform - the Raspberry Pi 4! This
is also going to include a lot of work on the idea of 'platforms' themselves
and inadequacies of the build system and such.

This leaves the Makefiles a bit of a mess but they'll be disappearing in
due course so it doesn't really matter.