kernel: require that `Platform` be 'static
Various kernel objects need to be parameterized over the `Platform` impl.
However, all kernel objects need to implement `Any` to support the current
downcasting implementation (this may change in the future). `Any` is of
course not implemented by types that aren't 'static, and the trait checker
reckons that even though we only use the `Platform` generic type to get
some associated types, the `Platform` also needs to be 'static.