drivers: start new echo driver to test 3 processes at the same time
This driver will be used for testing communication over a channel by simply
echoing back any messages sent to it and also used to build the "service"
advertisement system.

However, there appear to be a few problems in running 3 tasks at the same
time that weren't picked up with 2 tasks. So I guess we'll be fixing them
first :/