kernel: set new task to have a state of Running when switched to
This was never picked up because with two tasks (where the second one never
yielded), we never tried to switch from a task that wasn't dropped into by
drop_into_userspace (as opposed to being switched to by switch_to_next).
We fix this by actually changing the state of the new task to `Running`.