kernel: work around Mutex locking issue
The context switching works badly around Mutexes (and they're not really
the right primitive anyways), so for now replace them with an `UnsafeCell`
and not doing anything too stupid. This fixes the issues we had with
context switching to tasks multiple times.

In due course, this should be replaced by another primitive with safer
behavior (perhaps tied around the idea that we can pin a task to a specific
scheduler).