kernel_x86_64: don't leak kernel stack upon syscall
We were seeing page-faults when letting the kernel run for a long time with
a workload that actually switches between tasks and does a bunch of system
calls (I guess we've never actually tested this). Turns out the syscall
handler was leaking 8 bytes in the way it restored the kernel stack **on
every system call**. Whoops.

This now doesn't seem to leak any stack and runs for a really long time(TM).