kernel_x86_64: allow access of registers from interrupt handlers
This tidies up how we pass interrupt stack frames to Rust, and also saves
all the registers (instead of just caller-saved ones) and allows Rust to
access them. This will probably only be needed by the breakpoint handler,
but it's probably not that bad performance-wise (we can change it back
later if it proves to be slow), and is useful while we're trying to debug
this weird issue.

Interestingly, this changes the problem from presenting as a #PF into a #GP.