kernel_x86_64: store task user RSPs in per-CPU data
We used to save the user stack for a task on its kernel stack, which meant
we had to do some register gymnastics on syscalls, and also created a
complication when entering a task for the first time. Instead, we now store
it along with the kernel stack pointer in the task, and stick it in the
per-CPU data when a task is running.

Also document `do_drop_to_usermode`, as it was way cleverer than I realised
at the time and also not completely obvious how it's working at first glance.