kernel_x86_64: set FMASK and better docs
Use the new CpuFlags interface to create the `FMASK` for `syscall`, which
masks `rflags` when we enter the kernel. Also improve the documentation
for what the function actually does, including a simplified version of the
operation of `syscall`.

The FMASK it sets clears all the status flags (I'm not sure if we need to
do this, but this is intended to prevent miscompilations where the compiler
thinks it can assume a status flag is cleared and then it happens not to be,
depending on what userspace is doing). It also disables interrupts (this is
important because we faff about with the stacks, and we need interrupts to
be disabled while we do this), and clears the I/O privilege, nested task,
alignment check, and trap flags.