efiloader: align kernel stack and terminate stack frames
We were getting weird results from doing stack traces in the kernel, because
when we get to the kernel entry point, the frames carry on into the efiloader
and then crash because its stack is not longer mapped. This terminates
the stack frames by zeroing `rbp` before we enter the kernel.

We also correctly align the kernel stack (clearly Rust/LLVM doesn't actually
rely on it being aligned because we've never noticed it's not aligned, but
technically you need to).