hal_x86_64: simplify non-terminal flag handling
For some reason, we made life much more complicated than it needed to be
when handling the creation of parent tables. Instead, we can simply always
use the most permissive flags on parent tables, and then use the terminal
entries to actually control the permissions of a mapping.

Also rewrites a few comments that are clearly quite confused...