efiloader: make more sensible choices when no command line is supplied
We can't easily supply a command line on real hardware at the moment, so
we make more sensible choices here to allow us to test real hardware until
we work out how to do this.

Specifically, we assume a default kernel path, and also a basic GOP
framebuffer so the kernel can render some stuff to show that it's alive.