kernel_x86_64: factor out PCI access using the ECAM
We want to be able to access the PCI config space outside of enumeration,
so it seems simplest to simply pull it out into a new type. In the future,
we might even possibly support the ancient IO access mechanism, so we say
this is using the ECAM.

Also remove the AML routing table thingy since we don't need it for
enumeration.