[bundled] Update bundled OVMF and include license
I can't really remember where the old version of OVMF came from, so this
updates it to a version I've definitely built locally and included the
correct license for it. `efiloader` can now load the kernel with a default
OVMF_VARS file, which is nice, so this new version is the bog standard one.