hal: increase max number of memory entries
OVMF seems to create more than 64 memory entries now, somehow, so we need
to increase this to boot. Kind of a shame, but not really a big deal (the
boot info's memory map is now about a page long, which is not really a big
deal in the long run). If we care, we could probably do some coalescing
of the Pebble memory types in `efiloader`, but it probably just creates
extra work in the boot process for very little benefit.

In the future, we'll probably be able to reclaim the boot info memory
anyways.