kernel: send message to provider when a task subscribes to a service
At the moment, the message is empty and it just transfers the provider
end of the created channel to the provider. More info can always be added
to the body of the message (e.g. name of the subscribing task) if needed,
but we probably need to get ptah working first.