kernel: fix message delivery order through channels
Before, we were using a `Vec`, which effectively made message delivery FILO,
which is weird and could produce incorrect semantics if a task relies on
messages arriving in the correct order (and like, that's very plausible).
This moves channels over to storing messages in a `VecDeque`, which makes
it efficient to pop messages off the front, making delivery FIFO.

Also abstracts how we actually store messages inside a `ChannelEnd` by
introducing an `add_message` method to add a message to *this* ChannelEnd.