hal: increase number of memory entries again
We need to work out what the hell the UEFI seems to be doing with its
memory map, but we keep needing to increase this. We might need to start
doing some coalescensce of memory entries in the efiloader so we don't end
up with a 20M boot info at this rate.