Rename 'drivers' to 'user', defer to sub-makefile
We want to start building things that aren't really drivers per-se, so
make the directory for all user stuff, and then we can make sub-directories
when needed for things like USB host controller drivers or whatever. I think
that'll be cleaner than having a bunch of top-level directories, and also
allows us to share the target directory between all userspace applications,
which is a plus for build time and storage requirements.

Also build a sub-makefile, like we have for the kernel, to tidy a few things
up. In the near future, I think we'll be building a separate build system,
but this is fine for now.