efiloader: relax requirement for segment size to be multiple of page size
Various linking options don't actually make the segment size a multiple
of the page size like our method does, so we've relaxed it so that as long
as they don't overlap page ranges, it's fine, and we'll just zero some more
memory for them.

We still require the kernel to do it the way we do atm.