kernel: allow virtual address of MemoryObject to be `None` (WIP support)
This is to allow for MemoryObjects where the virtual address it is mapped
to doesn't matter, and so the mapping task can just pick where to map it
to. This just makes the first change (changing the type to an
Option<VirtualAddress>), so more work will be needed to actually allow
userspace to map it where it pleases etc.