platform_bus: rename Device to DeviceInfo, deal with handoff
The move to making handles actually-owned means device handoff from the
PlatformBus to a device driver a bit of a pain. We rename Device to
DeviceInfo to help suggest it's just some info about the device. We then
move this info to live in the state when the device is unclaimed, so we
can  move it out into the handoff message to be able to move the handles
out of PlatformBus.