kernel: pull user stacks into a TaskSlot abstraction
This changes the Address Space to allocate "task slots" instead of just
simply a user stack. This allows us to allocate more per-task things in
the address space (such as a Task Control Block + TLS space for each task).

Because of some stupid limitation of num_traits' detection of i128, this
also lowers the max number of tasks in an address space to 64 for now,
since this makes our lives easier now and we probably won't even need 64
for the foreseeable future (read probably ever).

This also moves to tracking the physical address of a Stack, which we'll
need if we want to write stuff to it later (and also it's good to track
in case we ever actually want to kill a task).