kernel: only allow buddy allocator to hand out blocks
This prevents a potential for physical memory leaks, and also makes the
separation of concerns in the physical allocator system better. The buddy
allocator can now only hand out entire blocks of physical memory (where
blocks are `2^i` frames for some `i`), and it's the responsibility of the
layer above it to work out how to actually satisfy an allocation. In the
future, this will allow us to do non-contiguous allocations, slab allocation
for specific sizes, etc.