efiloader: don't pass boot services memory to kernel if kludge enabled
This is the commit that caused me to need a break - what the hell is going
on here? We need to enable a kludge that keeps the UEFI boot services
identity mapped on some implementations after ExitBootServices because
broken - this probably means we shouldn't treat that memory as conventional
RAM on these broken systems. So we don't include it - I even had a naive
ideas that this would be the fix for our weird "my page tables become
completely corrupted at random points" bug, but oh no. Oh no.

This fundamentally breaks everything. Somehow, giving the kernel less memory
doesn't OOM it, but instead causes a bunch of memory corruption when we
switch between userspace tasks. Oh well, hopefully we'll get to the bottom
of this problem for real soon.