gpt: add rest of code before removing library
Irritatingly, someone fixed all the problems we couldn't be bothered to in
the `gpt` crate, so it's now much better to use, and we didn't realise. I
think it's best to switch to using that, but we'll keep our code around in
the history in case we want to pull it back out at any point.