tools: add our first EFI test (just a hello world)
efi_tests is a test suite to aims to check conformance of UEFI implementations.
hello_world is an incredibly simple test that literally just prints
"Hello, World!" to the UEFI console and halts, but has already proven
useful in confirming that we can make compiler_builtins provide mem* on
UEFI targets, instead of using rlibc as efiloader does.

The original intention of this test suite is to try and work out what
efiloader is doing wrong that causes OVMF to crash after we swap out the
page tables, but while building up some useful infrastructure.