efiloader: don't bubble errors - there's no point
If any of these things happen, `efiloader` should just crash. UEFI can't
do anything useful if we return to it in the vast majority of cases, so
we just don't bother. Instead, we'll just panic and make the panic screen
better. Some functions never actually even returned an error anyways.

I'm still undecided if "local" error bubbling makes sense (e.g. reporting
an error in load_kernel rather than load_segment so we can say it was the
kernel that failed to load.