efiloader: remove kludges, sanitise mappings a little
We had all sorts of odd stuff going on before, such as random bits of UEFI
code needing to be kept writable and boot services memory needing to be
made available or the kernel would randomly crash - this left a huge mess
in the loader as we thought UEFI was at fault. Now, UEFI is pretty terrible,
but I think this must be our fault somehow - it can't be that broken.

Remove all that stuff and the kernel still appears to load (without any
usermode tasks running however).