user: remove rlibc dependency from test_tls and simple_fb, add test_pf
By adding the compiler-builtins-mem feature during -Zbuild-std, we don't
need to depend on this extra library. This just removes it from the two
we're testing with at the moment, this should be done to the rest of them
too.

Also add a tiny little new testing task, which deliberately commits an
access violation to memory it shouldn't own. In some bizarre circumstances,
this is not correctly triggering our #PF handler, and instead triple faults,
which is very weird.