hal_x86_64: start testing page table implementation
Actually testing the paging implementation host-side is pretty difficult
as we do a lot of faffing with actual memory, but I want to at least be
able to test the algoriths such as the area-mapping stuff.

This is all part of the quest to find the bloody memory corruption bug.