efiloader: add a little extra space for the memory map, better comment
Our original comment might be a slight bit obnoxious: most UEFIs are broken
pieces of crap, but this is one of the more sorta reasonable reasons why.
Also explicitely add space for a few extra descriptors - we run the risk of
the size naturally being too close to the end of the frame to increase it
by a reasonable amount by aligning it up, crashing when we could just do
this instead.