efiloader: solve problem with needing boot services mapped!
We've solved it, and I'm not if the cause makes me look dumb, or UEFI even
worse - when UEFI loads your application, you obviously need a stack. It
could create this in LoaderData or wherever, but for whatever reason decides
that you don't need your stack at runtime or anything, and so puts it in a
memory region that we weren't mapping in the kernel page tables!? I can't be
arsed to put it back in to work out what region this is, but we had tracked
down that #PF to boot services code before, which just makes no bloody sense.

Anyway, now we've switched the kernel jump to a big block of assembly, we
don't touch the stack between switching to the kernel's set of page tables,
and switching to the new kernel stack, and so the UEFI stack disappearing
doesn't matter. Good riddance.