hal_x86_64: don't bother dynamically calculating GDT's limit
This seems overly complicated, and doesn't really serve any real purpose;
we allocate a fixed amount of memory for the TSSs anyway, so we might as
well tell the processor about all of it, and just leave the TSSs we don't
need empty. This will allow us to separate the creation of the GDT and
creation and loading of the TSS for each processor, which allows us to create
the GDT earlier, which makes our exception handlers actually work.