hal_x86_64,kernel_x86_46: create GDT earlier, without bootstrap TSS
This moves the creation of the GDT to earlier in the initialization sequence,
which is required for the installation of exception handlers to work correctly.
We missed this when we split out the exception handlers from the rest of
the IDT creation.

To load the GDT earlier, we separate out the creation of the bootstrap TSS,
which will happen later on, when we initialize the per-cpu data and topology
subsystem.

This commit also includes a test that the exception handlers are now in fact
working correctly.