various: commit various changes that have been sloshing about
Look, this isn't ideal. I haven't had time to work on Poplar at all recently,
and so various changes have been sloshing around. I couldn't even tell you
whether some of them are correct or not. But they're the current state of
the project, and so as a form of blank slate I'm committing them and we can
deal with my stupidity later if needed lmao.