kernel: get POC RISC-V kernel loading
Create crate for RISC-V kernel and add a proof-of-concept that can be loaded
by OpenSBI and prints a character to the serial port, just to prove it's
working. Because of the way RISC-V boots, I'm not yet sure if we'll need a
kind of 'pre-kernel' stage that loads the real kernel from an image or
whatever, but this works to start off.

Next step is to add functionality to build the kernel from `xtask`, and run
RISC-V QEMU