kernel_riscv: move entry-point to its own section
We were faffing around with introducing formatting machinery into the kernel,
which for some reason moves the entry point slightly (literally by a handful of
bytes; no idea why it feels the need to do that). Anyways, we can solve this
by putting the entry-point, which is the important bit, in its own section
and then manually placing it at the front of `.text` during linking.