hal: define `Size4KiB` on RISC-V
Various parts of `hal` basically just assume that 4K frames/pages are
supported, and so didn't compile on RISC-V. This is arguably not correct,
but since RISC-V does support this frame size, it's easiest just to define
it for now. I've left the other sizes undefined at this stage, as we don't
yet have a concrete plan for Sv39/48/57 support.