seed_riscv: build up a map of memory ranges
This adds a system whereby ranges of memory of particular types can be added,
with merging and intersecting regions automatically handled to build a
map of what memory is already occupied (e.g. the memory map could present
the whole of RAM as usable, and then specify bits that are already filled
with firmware, your code, etc.). This will help us build a memory allocator
later.