seed_riscv: mark device tree as reserved in memory ranges
This required `strict_provenance` to get the pointer's address, as apparently
now Rust doesn't allow that as a non-primitive cast? Ig in the past we've
just transmuted but I wanted to try the Cool New Way out, and I don't think
we need to do anything super weird except for enabling the feature and
not doing Weird Stuff with pointers that we've probably tried to avoid
anyways. We'll find out I guess...