seed_riscv: the simplest physical memory manager possible
Use the region map to construct an intrusive free list of usable memory,
from which we can build a super simple allocator. This can be extended to
handle frees if we need in the future, but I suspect we won't actually need
this functionality (definitely not for now, at least).