ptah: add support for `heapless` crate
We were going to move the Seed boot info to being serialized with `ptah`, but
I don't think this is the best approach now (due to high stack requirements
to build the type just to serialize it anyway). It's probably useful to keep
this stuff around anyways, now that we've built it.

### Wider context - how should we support third-party crates in `ptah`?
One of the problems we have in `ptah` is how we support third-party crates.
Very often, crates have support for `serde`, but obviously not `ptah`. Bc
of the orphan rules, we have two choices: the implementation either needs
to be in the crate, or in `ptah`.

There are effectively 3 options here:
- Maintain a fork of the crate with implementations added for `ptah`
- Implement a 'translation layer' between Serde implementations and `ptah`
- Add feature-gated implementations in `ptah` itself

A `serde` translation layer might be the best long-term solution, as it is
the only reasonable solution for third-party crates to implement. This would
probably be okay, as the Serde data model is pretty similar to `ptah`'s, but
there are some differences and I think the implementation would be slightly
messy in reality.

While the number of third-party crates needed is managable, I think we'll
go with the solution of adding feature-gated implementations to `ptah`
directly for now. They tend to be relatively small and gives us the greatest
freedom in terms of implementations and changes to the data model.