seed_riscv,kernel_riscv: pass boot info from Seed to kernel
The boot info is not yet very populated, but this constructs it and passes
it successfully to the kernel. Also adds the address of the FDT to it,
for platforms that have a device tree, so the kernel can access it.

We do unfortunately need to construct a default `BootInfo` on the stack
after all, so this did need us to increase the stack size a little.