Get Poplar building again on RISC-V and X64
We've disabled compiling user tasks on X64 for now because our custom
toolchain is completely broken. I think we should move away from requiring
it in the long-term because it's obvious it's not going to be upstreamed
at any point, and we don't have anywhere near enough time to maintain it
properly. Maybe explore an external `std` replacement to make user tasks
less painful?