`seed_uefi`: update to `acpi v0.24.0`
This moves away from the in-tree version of `uefi`, which has had all of
its changed upstreamed ages ago, to the upstream version. It turns out
they've made a bunch of cool changes which make the bootloader a fair
bit less painful!

Also fixes an old bug with the logger, where no ooutput was printed from
the graphics device selection. This is easily prevented, so let's do that.

Also makes some general improvements to code readability and updates some
comments to actually be accurate.