Remove the need for a custom Rust toolchain on x86_64 :)
We have long wanted to remove the requirement on the custom Rust toolchain,
because while it was a cool idea, we're never going to be able to upstream
it apparently, and it's a pain to maintain.

Instead, we're going to use an external `std` library and various big-brain
solutions to make it ergonomic to use. The first is a cool way to provide
a linker script, which I'm particularly happy with.