kernel_x86_64: rewrite `Tss` and `PerCpu` to not use `Pin` + correct layout
A new version of the compiler had reorganised the fields of `PerCpu`, which
broke everything. The underlying issue here is that we were not using the
correct representation for that struct, which is now fixed. However, we've
also removed all the usage of `Pin` - this is obviously a step backwards in
ideal soundness but they're so unergonomic to work with. Maybe come back and
improve the situation in the future?