`xtask`: customise user tasks from the config file
This is the start to a proper solution to the mess that is currently having
to independently specify which tasks to load from the Seed source lmao.
We can now specify which user tasks to load from the config, which will then
generate a config file in the image telling Seed what to load and from where.

A little more work will be needed to allow fully external tasks to be built,
because of workspaces. The format is also not great, mainly because it's
just TOML, so more complex tasks might need a nicer format so that it's not
too esoteric.