Remove in-tree version of OpenSBI
It turns out QEMU ships with an up-to-date version of OpenSBI by default,
so there isn't much point keeping it in-tree. I don't know why we thought
this was necessary in the first place.

This does keep the code for customising it, to make that easy in the future
if we need to.

Closes #90 - this is unnecessary if no in-tree version