`lib/gpt`: start a very basic library for dealing with GPT from `no_std`
Just has a GUID definition and some structs for the header etc. for now.
This was necessary because the `gpt` library we use in the xtask doesn't
support `no_std`, and looks not-at-all-simple to convert to supporting it.
We don't need very much support at the moment, so I thought we'd just
write a very basic one of our own.