`xtask`,`opensbi`: add `opensbi` submodule and `xtask` command to build it
We had this in-tree a while ago and got rid of it because QEMU ships with
its own version. However, now we're building for hardware, we need a way
to supply it, and it seems sensible to have a nice `xtask` command to build
it. In the future this can be extended to know how to build OpenSBI for
each platform.