`seed/d1_boot0`: add an amazingly simple test program for the D1
This is just intended to prove that the D1 + FEL etc actually does work,
because naively uploading OpenSBI and running it does not produce output
(probably because it doesn't know where the UART is without a device tree
lmao). It's named `d1_boot0` because it will become the boot-shim (known as
the SPL in U-Boot terminology, or `boot0` in the D1 bootloader) for both
FEL booting and booting from a persistent medium.