Add a `memory` node to the MangoPi MQ-Pro device tree
I'm not sure why the original device tree doesn't include this - a mystery
for the future because Seed needs one to boot, and we can easily just add
it here?

There are actually multiple models of the MQ-Pro that have different amounts
of RAM, so we could support that in the future.