`xtask`: gracefully handle invalid UTF-8 from the UART
The UART spuriously produces invalid UTF-8, presumably because of corrupted
bytes (likely due to the slightly dodgy situation with the wiring). We'll
resolder it when we have access to an iron, but for now we (and probably
for the long-term) it's probably a good idea to just accept dodgy input,
rather than hard-erroring.