`hal_riscv`: rename references to top-level page table in paging impl
This is the first step in making the paging impl generic over how many levels
the tables have. I'm thinking we'll be able to add a generic for the top-level
to `PageTableImpl`, and then somehow specialise each method to how many
levels need to be traversed, but we may discover that's a pain in the ass
and either a) do it at runtime with an enum or something or b) make separate
types for each paging scheme and make it the user's problem to be generic
over them (this also makes it my problem but slightly later in time).