`hal_riscv`: build support for the `Sv39` paging mode
The D1 only supports the `Sv39` paging mode, so we need to be able to support
both (we could also move to just supporting Sv39 and ditching Sv48 even on
systems that support it, but we've done this for now so unless it becomes
annoying to maintain let's leave it). This wasn't actually too difficult,
but does duplicate a lot of code. In the future, we could possibly find a
way to deduplicate this by making it generic across the top-level table.