Provide memory layout and other constants through `hal_riscv::platform` module
I'm not sure if this will turn out to be the right solution in the long-term,
but for now seems alright. We effectively need to change various things,
such as the layout of the physical and memory address spaces, and the paging
scheme to use, depending on the platform we're building for. This happens
by setting a feature in the built crate, which is then passed along to
`hal_riscv`, which then re-exports the correct module as `hal_riscv::platform`.